Refactor: Move MacroBrowser.js to macros/engine directory and update import paths
